Added an existence check for PendingMember object before removing it. Volunteers added directly from website using add volunteer form doesn't have a PendingMember object for them.

+++ cleansweep/plugins/volunteers/views.py
@@ -95,7 +95,8 @@ def profile(id, hash):
             from ..audit.models import Audit
             Audit.query.filter_by(person_id=m.id).delete()
             Audit.query.filter_by(user_id=m.id).delete()
-            db.session.delete(pending_member)
+            if pending_member is not None:
+                db.session.delete(pending_member)
             db.session.delete(m)
             db.session.commit()
             signals.delete_volunteer.send(m, place=place)
